refactoring : refactored UsersPagination and hooks it uses.
