+++ src/modules/users/hooks/helpers/usePagesCount.js
@@ -1,0 +1,23 @@
+import { useMemo } from 'react';
+import useUsersQueryParams from '@/modules/users/hooks/state/useUsersQueryParams';
+import useUsers from '@/modules/users/hooks/api/useUsers';
+import getPagesCount from '@/modules/users/helpers/getPagesCount';
+
+const usePagesCount = () => {
+  const { usersQueryCount } = useUsersQueryParams();
+  const { data, isLoading } = useUsers();
+
+  const usersTotalCount = data?.totalCount;
+
+  const pagesCount = useMemo(() => {
+    if (!isLoading && usersTotalCount) {
+      return getPagesCount(usersTotalCount, usersQueryCount);
+    }
+
+    return 1;
+  }, [isLoading, usersQueryCount, usersTotalCount]);
+
+  return pagesCount;
+};
+
+export default usePagesCount;
